backend/requirements.txt
openai>=1.3.0
chromadb>=0.4.18
numpy>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
python-docx>=1.1.2
pypdf>=5.0.0
openpyxl>=3.1.5
xlrd>=2.0.1
beautifulsoup4>=4.12.3
pytesseract>=0.3.10
pdf2image>=1.17.0
pillow>=11.0.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
requests>=2.31.0
pydantic>=2.0.0
